requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.2
django-filter==0.10.0
django-guardian==1.3
djangorestframework==3.1.3
duckduckgo==0.1
duckduckgo2==0.242
gunicorn==19.3.0
Markdown==2.6.2
psycopg2==2.6
requests==2.7.0
six==1.9.0
static3==0.6.1
wheel==0.24.0
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.2
django-filter==0.10.0
django-guardian==1.3
djangorestframework==3.1.3
duckduckgo==0.1
duckduckgo2==0.242
gunicorn==19.3.0
Markdown==2.6.2
poster==0.8.1
psycopg2==2.6
requests==2.7.0
six==1.9.0
static3==0.6.1
Unirest==1.1.7
wheel==0.24.0
